first cut of a rough weather service

+++ src/Core/Models/FeelsLike.cs
@@ -5,14 +5,14 @@ namespace Core.Models;
 public class FeelsLike
 {
     [JsonPropertyName("day")]
-    public double? Day { get; set; }
+    public decimal? Day { get; set; }
 
     [JsonPropertyName("night")]
-    public double? Night { get; set; }
+    public decimal? Night { get; set; }
 
     [JsonPropertyName("eve")]
-    public double? Eve { get; set; }
+    public decimal? Eve { get; set; }
 
     [JsonPropertyName("morn")]
-    public double? Morn { get; set; }
+    public decimal? Morn { get; set; }
 }
