+++ src/Core/Models/Minutely.cs
@@ -5,7 +5,7 @@ namespace Core.Models;
 public class Minutely
 {
     [JsonPropertyName("dt")]
-    public int? Dt { get; set; }
+    public long? Dt { get; set; }
 
     [JsonPropertyName("precipitation")]
     public int? Precipitation { get; set; }
